refactor(runtime): split workers into modules
